refactor : message 도메인의 패키지 구조 변경
