+++ src/main/java/com/example/githubprconsumer/message/application/MessageService.java
@@ -1,5 +1,8 @@
-package com.example.githubprconsumer.message;
+package com.example.githubprconsumer.message.application;
 
+import com.example.githubprconsumer.message.application.dto.GithubPRResponse;
+import com.example.githubprconsumer.message.domain.Message;
+import com.example.githubprconsumer.message.domain.MessageRepository;
 import com.example.githubprconsumer.messenger.MessengerException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
