+++ src/main/java/com/example/githubprconsumer/message/domain/TemplatePlaceholder.java
@@ -1,4 +1,4 @@
-package com.example.githubprconsumer.message;
+package com.example.githubprconsumer.message.domain;
 
 import lombok.Getter;
