+++ src/test/java/com/example/githubprconsumer/message/MessageServiceTest.java
@@ -1,5 +1,9 @@
 package com.example.githubprconsumer.message;
 
+import com.example.githubprconsumer.message.application.MessageService;
+import com.example.githubprconsumer.message.application.dto.GithubPRResponse;
+import com.example.githubprconsumer.message.domain.Message;
+import com.example.githubprconsumer.message.domain.MessageRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
